requirements.txt
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-pptx>=0.6.21
pillow>=10.0.0
plotly>=5.18.0
pandas>=2.1.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-multipart>=0.0.6
psutil>=5.9.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0 